+++ source/src/sfincs_momentum.f90
@@ -10,10 +10,10 @@ contains
    !
    ! Computes fluxes over subgrid u and v points
    !
-   integer   :: count0
-   integer   :: count1
-   integer   :: count_rate
-   integer   :: count_max
+   integer*8 :: count0
+   integer*8 :: count1
+   integer*8 :: count_rate
+   integer*8 :: count_max
    real      :: tloop
    !
    real*4    :: dt
