+++ source/src/sfincs_output.f90
@@ -87,10 +87,10 @@ module sfincs_output
    !
    implicit none
    !
-   integer  :: count0
-   integer  :: count1
-   integer  :: count_rate
-   integer  :: count_max
+   integer*8 :: count0
+   integer*8 :: count1
+   integer*8 :: count_rate
+   integer*8 :: count_max
    real     :: tloop
    !
    logical  :: write_map
